moved the text to Rmarkdown

+++ project_support.r
@@ -32,16 +32,25 @@ modal <- function(data){
   return(mode)
 }
 
-dir_init <- function(path, verbose=FALSE){
+dir_init <- function(path, verbose = FALSE, overwrite = TRUE){
   if(substr(path, 1, 2)!='./') stop('path argument must be formatted
     with "./" at beginning')
-  contents <- dir(path)
-  if(verbose){
-    if(length(contents)==0) print(paste('folder ', path, ' created.', sep=""))
-    if(length(contents)>0) print(paste('folder ', path, ' wiped of ', length(contents), ' files/folders.', sep=""))
+  contents <- dir(path, recursive=TRUE)
+  if(dir.exists(path)){
+    if(overwrite){
+      if(verbose){
+        if(length(contents)==0) print(paste('folder ', path, ' created.', sep=""))
+        if(length(contents)>0) print(paste('folder ', path, ' wiped of ', length(contents), ' files/folders.', sep=""))
+      }
+      if(dir.exists(path)) unlink(path, recursive=TRUE)
+      dir.create(path)
+    }
+  } else {
+    if(verbose){
+      print(paste('folder ', path, ' created.', sep=""))
+    }
+    dir.create(path)
   }
-  if(dir.exists(path)) unlink(path, recursive=TRUE)
-  dir.create(path)
 }
 
 sapply_pb <- function(X, FUN, ...){
